requirements.txt
openai>=1.42.0
python-dotenv>=1.0.1
requests>=2.32.3
PyMuPDF>=1.24.9
duckduckgo-search>=5.3.1
rank-bm25>=0.2.2
tqdm>=4.66.4